Updated swap keys trigger timer. Added preference static class to hold important string variables

+++ Endless-Geo-Dash/Assets/Scripts/UI/ScoreDisplay.cs
@@ -35,7 +35,7 @@ namespace GPEJ.UI
         {
             if (Time.timeScale != 0f)
             {
-                ui_text.text = runtime_data.distance.ToString();
+                ui_text.text = ((int)(runtime_data.distance)).ToString();
             }
         }
     }
